AI-agent/campus_ai_secretary/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
qdrant-client>=1.7.0
redis>=5.0.0
dashscope>=1.14.0
python-dotenv>=1.0.0
loguru>=0.7.0
python-dateutil>=2.8.0
httpx>=0.25.0
orjson>=3.9.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0
